fix: drone position 함수 분리
